Move the XHR callback to a separate file
