+++ lib/send-request.js
@@ -13,21 +13,21 @@ module.exports = function(location, options, callback) {
   request.onreadystatechange = function() {
     if (request.readyState === 4) {
       if (request.status === 200) {
-        callback(request.responseText, request)
+        callback(request.responseText, request, location)
       }
       else {
-        callback(null, request)
+        callback(null, request, location)
       }
     }
   }
 
   request.onerror = function(e) {
     console.log(e)
-    callback(null, request)
+    callback(null, request, location)
   }
 
   request.ontimeout = function() {
-    callback(null, request)
+    callback(null, request, location)
   }
 
   // Prepare the request payload for forms, if available
